#28 refactoring: share code  in AbstractXpp3Generator
sourceTracker is also optional
